// MAIN/PCES.PaymentPlatform.WebUI/PCES.PaymentPlatform.ReactApp/src/store/auth/types.ts
import { IUser } from '../../models/IUser';
import { IError } from '../../models/IError';

export enum AuthActionTypes {
  IS_SUBMITTING = '@@auth/IS_SUBMITTING',
  LOGIN_SUCCESS = '@@auth/LOGIN_SUCCESS',
  LOGIN_ERROR = '@@auth/LOGIN_ERROR',
  REQUEST_LOGIN_SUBMIT = '@@auth/REQUEST_LOGIN_SUBMIT',
  REQUEST_REGISTER_SUBMIT = '@@auth/REQUEST_REGISTER_SUBMIT',
  REGISTER_SUCCESS = '@@auth/REGISTER_SUCCESS',
  REGISTER_ERROR = '@@auth/REGISTER_ERROR',
  RESET_ERROR_STATE = '@@auth/RESET_ERROR_STATE',
  LOG_SINGLE_ERROR = '@@auth/LOG_SINGLE_ERROR',
  SET_USER = '@@auth/SET_USER',
  DESTROY_STATES = '@@auth/DESTROY_STATES',
}

export interface AuthState {
  readonly isAuthenticated: boolean;
  readonly isSubmitting: boolean;
  readonly error: IError | {};
  readonly success: boolean;
  readonly user: IUser | {};
  readonly singleError: string;
}